+++ LineSDK/LineSDKObjCInterfaceTests/LineSDKModelInterfaceTests.m
@@ -55,7 +55,6 @@
     XCTAssertNil(token.value);
     XCTAssertNil(token.createdAt);
     XCTAssertNil(token.IDToken);
-    XCTAssertNil(token.refreshToken);
     XCTAssertNil(token.permissions);
     XCTAssertNil(token.expiresAt);
     XCTAssertNil(token.json);
@@ -118,6 +117,7 @@
     XCTAssertFalse(manager.isSetupFinished);
     XCTAssertFalse(manager.isAuthorized);
     XCTAssertFalse(manager.isAuthorizing);
+
     [manager setupWithChannelID:@"" universalLinkURL:nil];
     [manager loginWithPermissions:nil
                  inViewController:nil
@@ -139,6 +139,15 @@
     XCTAssertFalse(opened);
     
     XCTAssertNotNil([LineSDKLoginManager sharedManager]);
+}
+
+- (void)testLoginManagerLangSettingInterface {
+    LineSDKLoginManager *manager = [LineSDKLoginManager sharedManager];
+    XCTAssertNil(manager.preferredWebPageLanguage);
+    [manager setPreferredWebPageLanguage:@"zh-Hans"];
+    XCTAssertEqual(manager.preferredWebPageLanguage, @"zh-Hans");
+    manager.preferredWebPageLanguage = nil;
+    XCTAssertNil(manager.preferredWebPageLanguage);
 }
 
 - (void)testHexColorInterface {
@@ -255,6 +264,7 @@
     XCTAssertNil(payload.name);
     XCTAssertNil(payload.picture);
     XCTAssertNil(payload.email);
+    XCTAssertNil(payload.amr);
 }
 
 - (void)testLoginButtonInterface {